requirements.txt
pytrends>=4.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.30.0
anthropic>=0.25.0
pyyaml>=6.0
rich>=13.0.0
